Heavy Commit!
Home-Page UI are now more flexible.
BLoC was implemented in the last merge.
The file structure has been changed for better organization.
Navigation system is working now and ready for handling page change events.
The Position of the bottom card has been fixed so that it is always at the bottom.
Intro and splash screens now use the same color schemes as the home screen.
Font styles and colors have been added to the constants folder as needed.

NOTE: The other platforms folders have been removed to reduce problems when cloning the project,
to create them use ``flutter create . --platforms <The Platform>``

Fixed "Failed to transform core-for-system-modules.jar to match attributes" by upgrading android from '8.1.0' to '8.7.1'