Fix 'Juz', 'audio_player' Bugs & Add some touches.
Selecting something and then moving to another screen will not reset the selection.
This will prevent the user from selecting the last selection again.

From a user feedback the was in Juz screen but the same problem appears in the Surah (audio tab).

FIXED: There was a typo in 'audioPlayerDismissibleHintText' but it got fixed in this commit ;)
NEW: When playing an audio file, the Surah screen will automatically go to Index 1 (AudioScreen) When the user navigate to it.
NOTE: TODO at aea704f5b06949f40d7c4e9253dc210bfe7d558e had been completed in the previous commit.